% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cool_math.R
\name{cool_math}
\alias{cool_math}
\title{Cool math function}
\usage{
cool_math(x, y)
}
\arguments{
\item{x}{a numeric value of x}

\item{y}{a numeric value of y}
}
\value{
a numeric value that is the sum of the squares of x and y
}
\description{
A function that squares x and squares y then adds them
}
\examples{
cool_math(4, 5)
}
