% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duck_goose.R
\name{duck_goose}
\alias{duck_goose}
\title{Duck and geese sentence building function}
\usage{
duck_goose(adj, ducks, geese)
}
\arguments{
\item{adj}{A character string adjective you would like to use to describe your avian friends}

\item{ducks}{A number that tells how many ducks you have}

\item{geese}{A number that tells how many geese you have}
}
\value{
A sentence that lists total ducks and geese with your selected adjective
}
\description{
A little function that adds how many ducks and geese you have and returns a sentence with the output
}
\examples{
duck_goose(adj = "funny", ducks = 4, geese = 8)
"I have 12 funny ducks and geese!"
}
