% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_me.R
\name{help_me}
\alias{help_me}
\title{Help me sentence builder}
\usage{
help_me(task, hours)
}
\arguments{
\item{hours}{a numeric value of the number of hours you've been doing the thing}

\item{a}{character string for the task or activity you're doing}
}
\value{
a character string sentence that expresses your despair
}
\description{
Help me sentence builder
}
\examples{
help_me("data science", 10)
}
